pyproject.toml
[tool.poetry]
name = "back-end"
version = "0.1.0"
description = "Desafio Hyperativa"
authors = ["Rafael Sales <rafael.sales@gmail.com>"]

[tool.poetry.dependencies]
python = "3.8.9"
Django = "3.1.9"
djangorestframework = "^3.12.4"
djangorestframework-simplejwt = "^4.6.0"
Pillow = "^8.2.0"
drf-yasg = "^1.20.0"
psycopg2-binary = "^2.8.6"
django-configurations = "^2.2"
django-cors-headers = "^3.7.0"
gunicorn = "^20.1.0"
django-dotenv = "^1.4.2"
python-dotenv = "^0.17.1"
whitenoise = "^5.2.0"
dj-database-url = "^0.5.0"
django-utils-six = "^2.0"
django-searchable-encrypted-fields = "^0.1.9"
drf-api-logger = "^1.0.6"

[tool.poetry.dev-dependencies]
jupyterlab = "^3.0.14"
pydotplus = "^2.0.2"
coverage = "^5.5"
pycodestyle = "^2.7.0"
bandit = "^1.7.0"
yapf = "^0.31.0"
selenium = "^3.141.0"
Faker = "^8.1.2"
sphinx-rtd-theme = "^0.5.2"
Sphinx = "^3.5.4"
pytest-django = "^4.2.0"
pytest = "^6.2.4"
pandas = "^1.2.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
